template method products
